requirements.txt
streamlit==1.47.0
authlib==1.3.0
requests-oauthlib==1.3.1
cryptography==41.0.7 